requirements.txt
flask==3.1.0
python-dotenv==1.1.0
requests==2.32.3
scikit-learn==1.6.1
python-docx==1.1.2
pdfplumber==0.11.6
werkzeug==3.1.3 